services/cv_engine/requirements.txt
ultralytics
opencv-python
deep-sort-realtime
kafka-python
numpy